refactor(cli): :recycle: rename deamon to daemon
