+++ backend/tests/test_cli_daemon.py
